requirements/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile requirements/requirements.in
#
--extra-index-url https://pypi.ngc.nvidia.com
--trusted-host pypi.ngc.nvidia.com

anyio==3.7.1
    # via starlette
asgiref==3.7.2
    # via django
click==8.1.6
    # via uvicorn
colorama==0.4.6
    # via click
django==4.2.5
    # via -r requirements/requirements.in
ecdsa==0.18.0
    # via python-jose
exceptiongroup==1.1.3
    # via anyio
fastapi==0.101.1
    # via
    #   -r requirements/requirements.in
    #   fastapi-utils
fastapi-utils==0.2.1
    # via -r requirements/requirements.in
greenlet==2.0.2
    # via sqlalchemy
h11==0.14.0
    # via uvicorn
idna==3.4
    # via anyio
numpy==1.25.2
    # via opencv-python
opencv-python==4.8.0.76
    # via -r requirements/requirements.in
packaging==23.1
    # via pytesseract
passlib==1.7.4
    # via -r requirements/requirements.in
pillow==10.0.0
    # via pytesseract
pyasn1==0.5.0
    # via
    #   python-jose
    #   rsa
pydantic==1.10.12
    # via
    #   fastapi
    #   fastapi-utils
pytesseract==0.3.10
    # via -r requirements/requirements.in
python-jose==3.3.0
    # via -r requirements/requirements.in
rsa==4.9
    # via python-jose
six==1.16.0
    # via ecdsa
sniffio==1.3.0
    # via anyio
sqlalchemy==1.4.49
    # via fastapi-utils
sqlparse==0.4.4
    # via django
starlette==0.27.0
    # via fastapi
typing-extensions==4.7.1
    # via
    #   asgiref
    #   fastapi
    #   pydantic
    #   starlette
    #   uvicorn
tzdata==2023.3
    # via django
uvicorn==0.23.2
    # via -r requirements/requirements.in
